+++ error-prone-contrib/src/test/resources/tech/picnic/errorprone/refasterrules/ImmutableEnumSetRulesTestOutput.java
@@ -1,23 +1,31 @@
 package tech.picnic.errorprone.refasterrules;
 
+import static com.google.common.collect.ImmutableSet.toImmutableSet;
+import static com.google.common.collect.Sets.toImmutableEnumSet;
+
+import com.google.common.collect.BoundType;
 import com.google.common.collect.ImmutableSet;
+import com.google.common.collect.Iterables;
 import com.google.common.collect.Sets;
 import java.math.RoundingMode;
 import java.util.Arrays;
 import java.util.EnumSet;
+import java.util.stream.Stream;
 import tech.picnic.errorprone.refaster.test.RefasterRuleCollectionTestCase;
 
 final class ImmutableEnumSetRulesTest implements RefasterRuleCollectionTestCase {
   @Override
   public ImmutableSet<Object> elidedTypesAndStaticImports() {
-    return ImmutableSet.of(EnumSet.class);
+    return ImmutableSet.of(EnumSet.class, toImmutableSet());
   }
 
-  ImmutableSet<RoundingMode> testSetsImmutableEnumSetIterable() {
-    return Sets.immutableEnumSet(EnumSet.range(RoundingMode.UP, RoundingMode.UNNECESSARY));
+  ImmutableSet<ImmutableSet<RoundingMode>> testSetsImmutableEnumSetIterable() {
+    return ImmutableSet.of(
+        Sets.immutableEnumSet(Iterables.cycle(RoundingMode.UP)),
+        Sets.immutableEnumSet(EnumSet.allOf(RoundingMode.class)));
   }
 
-  ImmutableSet<RoundingMode> testSetsImmutableEnumSetIterableArray() {
+  ImmutableSet<RoundingMode> testSetsImmutableEnumSetArraysAsList() {
     return Sets.immutableEnumSet(Arrays.asList(RoundingMode.values()));
   }
 
@@ -72,7 +80,7 @@ final class ImmutableEnumSetRulesTest implements RefasterRuleCollectionTestCase 
         RoundingMode.HALF_EVEN);
   }
 
-  ImmutableSet<RoundingMode> testImmutableEnumSetVarArgs() {
+  ImmutableSet<RoundingMode> testSetsImmutableEnumSetVarArgs() {
     return Sets.immutableEnumSet(
         RoundingMode.UP,
         RoundingMode.DOWN,
@@ -80,5 +88,9 @@ final class ImmutableEnumSetRulesTest implements RefasterRuleCollectionTestCase 
         RoundingMode.FLOOR,
         RoundingMode.UNNECESSARY,
         RoundingMode.HALF_EVEN);
+  }
+
+  ImmutableSet<BoundType> testStreamToImmutableEnumSet() {
+    return Stream.of(BoundType.OPEN).collect(toImmutableEnumSet());
   }
 }
